# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic API"
  description: "This is an API"
  version: "1.0.0"
host: "autoshop-api-j2nm.onrender.com"
schemes:
  - "https"
  - "http"
consumes:
  - "application/json" # My api receives JSON data
produces:
  - "application/json" # My api produces JSON data
securityDefinitions: # Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Endpoint to generate a new mechanic"
      description: "Endpoint to create a new mechanic in AutoShop system, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new mechanic."
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"
      responses:
        201:
          description: "Mechanic Creation Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              name: "Test Mech"
              email: "test@email.com"
              phone: "123-456-7890"
              salary: 99999.99

    get:
      tags:
        - Mechanics
      summary: "Returns all Mechanics"
      description: "Endpoint to receive an array of mechanics objects"
      responses:
        200:
          description: "Successfully retrieved all mechanics"
          schema:
            $ref: "#/definitions/MechanicsResponse"

    put:
      tags:
        - Mechanics
      summary: "Update mechanic endpoint"
      description: "Endpoint to update a mechanic"
      parameters:
        - in: "body"
          name: "body"
          description: "Update Mech"
          required: true
          schema:
            $ref: "/definitions/MechanicInput"
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 2
              name: "Test Mechanic"
              email: "test@email.com"
              phone: "123-456-9999"
              salary: 99999.99

    delete:
      tags:
        - Mechanics
      summary: "Delete a Mechanic"
      description: "Delete a Mechanic from DB"
      responses:
        200:
          schema:
            type: object
            properties:
              messsage:
                type: string
...
definitions:
...
  MechanicInput:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: float
      
    required:
      - name
      - email
      - phone
      - salary
  MechanicResponse:
    type: "object"
    properties:
      id:
        type: "int"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      salary:
        type: float
      
    MechanicsResponse:
      type: "array"
      items:
        type: "object"
        properties:
          id:
            type: "int"
          name:
            type: "string"
          email:
            type: "string"
          phone:
            type: "string"
          salary:
            type: float
...